skills/database/dms/alibabacloud-data-agent-skill/scripts/requirements.txt
# Data Agent CLI 依赖
# 推荐使用虚拟环境安装:
# python3 -m venv venv && source venv/bin/activate
# pip install -r requirements.txt

# 阿里云 SDK
alibabacloud-tea-openapi>=0.3.0
alibabacloud-dms20250414>=1.0.0
alibabacloud-dms-enterprise20181101>=1.0.0
alibabacloud-openapi-util>=0.2.0

# HTTP 客户端
requests>=2.28.0
aiohttp>=3.9.0

# 环境变量
python-dotenv>=1.0.0
